compile ze_geometry with float

+++ src/ransac_relative_pose.cpp
@@ -89,8 +89,8 @@ bool RansacRelativePose::solveRelativePose(
            << ", #inliers = " << ransac.inliers_.size();
 
   // Process results.
-  Matrix3 R = ransac.model_coefficients_.leftCols<3>();
-  Vector3 t = ransac.model_coefficients_.rightCols<1>();
+  Matrix3 R = ransac.model_coefficients_.leftCols<3>().cast<FloatType>();
+  Vector3 t = ransac.model_coefficients_.rightCols<1>().cast<FloatType>();
   T_cur_ref = Transformation(Quaternion(R), t);
   result_probability_ = ransac.probability_;
   num_iterations_ = ransac.iterations_;
@@ -109,7 +109,7 @@ bool RansacRelativePose::solveTranslationOnly(
   //! @todo: Unify all the repetitive code.
   using Problem = opengv::sac_problems::relative_pose::TranslationOnlySacProblem;
   using Adapter = opengv::relative_pose::CentralRelativeAdapter;
-  Adapter adapter(f_cur, f_ref, T_cur_ref.getRotationMatrix());
+  Adapter adapter(f_cur, f_ref, T_cur_ref.getRotationMatrix().cast<double>());
   boost::shared_ptr<Problem> problem(new Problem(adapter));
   opengv::sac::Ransac<Problem> ransac;
   ransac.sac_model_ = problem;
@@ -128,8 +128,8 @@ bool RansacRelativePose::solveTranslationOnly(
            << ", #inliers = " << ransac.inliers_.size();
 
   // Process results.
-  Matrix3 R = ransac.model_coefficients_.leftCols<3>();
-  Vector3 t = ransac.model_coefficients_.rightCols<1>();
+  Matrix3 R = ransac.model_coefficients_.leftCols<3>().cast<FloatType>();
+  Vector3 t = ransac.model_coefficients_.rightCols<1>().cast<FloatType>();
   T_cur_ref = Transformation(Quaternion(R), t);
   result_probability_ = ransac.probability_;
   num_iterations_ = ransac.iterations_;
@@ -166,8 +166,8 @@ bool RansacRelativePose::solveRotationOnly(
            << ", #inliers = " << ransac.inliers_.size();
 
   // Process results.
-  Matrix3 R = ransac.model_coefficients_.leftCols<3>();
-  Vector3 t = ransac.model_coefficients_.rightCols<1>();
+  Matrix3 R = ransac.model_coefficients_.leftCols<3>().cast<FloatType>();
+  Vector3 t = ransac.model_coefficients_.rightCols<1>().cast<FloatType>();
   T_cur_ref = Transformation(Quaternion(R), t);
 
   result_probability_ = ransac.probability_;
